change dot-avatar for dot-gravatar and fix tests
